Move omit checks from the formatter to the engine
