+++ tests/phpunit/integration/DiscordRCFeedFormatterTest.php
@@ -80,71 +80,6 @@ class DiscordRCFeedFormatterTest extends MediaWikiIntegrationTestCase {
 		);
 	}
 
-	public static function providerGetLineOmittance() {
-		return [
-			[
-				[
-					'omit_namespaces' => [ NS_TALK ],
-				],
-				[
-					'rc_namespace' => NS_TALK,
-				],
-				'should omit the given namespaces'
-			],
-			[
-				[],
-				[
-					'rc_type' => RC_CATEGORIZE,
-				],
-				'should omit RC_CATEGORIZE change always'
-			],
-			[
-				[
-					'omit_log_types' => 'patrol',
-				],
-				[
-					'rc_type' => RC_LOG,
-					'rc_log_type' => 'patrol',
-					'rc_log_action' => 'patrol',
-				],
-				'should omit the given log types'
-			],
-			[
-				[
-					'omit_log_actions' => 'patrol/patrol-auto',
-				],
-				[
-					'rc_type' => RC_LOG,
-					'rc_log_type' => 'patrol',
-					'rc_log_action' => 'patrol-auto',
-				],
-				'should omit the given log action'
-			],
-		];
-	}
-
-	/**
-	 * @dataProvider providerGetLineOmittance
-	 * @covers \MediaWiki\Extension\DiscordRCFeed\DiscordRCFeedFormatter::getLine
-	 */
-	public function testOmitChanges( array $feed, array $attribs, string $message = '' ) {
-		// Provide mandatory parameters if not given
-		$attribs = array_replace_recursive( [
-			'rc_namespace' => NS_MAIN,
-			'rc_title' => 'Test page',
-			'rc_type' => RC_EDIT,
-		], $attribs );
-		MediaWikiServices::initializeParameters( $feed, [
-			'style' => 'embed',
-			'user_tools' => [],
-			'page_tools' => [],
-		], [ 'omit_types' => [ RC_CATEGORIZE ] ] );
-
-		$rc = self::makeRecentChange( $attribs );
-		$rt = $this->formatter->getLine( $feed, $rc, '' );
-		$this->assertNull( $rt, $message );
-	}
-
 	public static function providerGetDescription() {
 		return [
 			[
